backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "synopticspread"
version = "0.1.0"
description = "Track NWP meteorological model divergence"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "herbie-data>=2024.9.0",
    "cdsapi>=0.7.3",
    "xarray>=2024.11.0",
    "cfgrib>=0.9.15",
    "metpy>=1.6.0",
    "zarr>=2.18.0",
    "apscheduler>=3.11.0",
    "numpy>=2.1.0",
    "pandas>=2.2.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.35.0",
    "ruff>=0.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
